feat: 메인(Home) 화면에서 db.json의 데이터를 가지고 오도록 수정
