+++ frontend/src/App.jsx
@@ -1,7 +1,6 @@
 import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
 import "./App.css";
-// import Home from "./pages/Home.jsx";
-import HomeAPI from "./pages/HomeAPI.jsx";
+import Home from "./pages/Home.jsx";
 import SearchPage from "./pages/SearchPage.jsx";
 import BookInfo from "./pages/BookInfo.jsx";
 import Footer from "./components/Footer.jsx";
@@ -12,8 +11,7 @@ function App() {
   return (
     <Router>
       <Routes>
-        {/* <Route path="/" element={<Home />} /> */}
-        <Route path="/" element={<HomeAPI />} />
+        <Route path="/" element={<Home />} />
         <Route path="/search" element={<SearchPage />} />
         <Route path="/bookinfo/:isbn" element={<BookInfo />} />
         <Route path="/success" element={<SuccessPage />} />
