+++ frontend/src/components/BookCarousel.jsx
@@ -1,61 +1,49 @@
 import React, { useEffect, useState } from "react";
-import "../styles/BookCarousel.css"; // CSS 파일을 import합니다.
+// import axios from "axios";
+import "../styles/BookCarousel.css";
+import defaultThumbnail from "../assets/default-book-image.png";
+import { fetchBooks } from "../api/books";
 import iconLeft from "../assets/icon/png/angle-left.png";
 import iconRight from "../assets/icon/png/angle-right.png";
 
-function BookCarousel({ query, title }) {
-  const [booksData, setBooksData] = useState([]);
-  const [currentIndex, setCurrentIndex] = useState(0); // ★ 추가: 현재 보여줄 책 인덱스
-  const apiKey = import.meta.env.VITE_KAKAO_REST_API_KEY;
+function BookCarousel({ category, title }) {
+  const [booksData, setBooksData] = useState([]); // 배열로 선언
+  const [currentIndex, setCurrentIndex] = useState(0);
 
   useEffect(() => {
-    const fetchBooks = async () => {
-      try {
-        const response = await fetch(
-          `https://dapi.kakao.com/v3/search/book?target=title&query=${encodeURIComponent(query)}&sort=latest&size=30`,
-          { headers: { Authorization: `KakaoAK ${apiKey}` } }
-        );
-        if (!response.ok) {
-          throw new Error(`HTTP error! status: ${response.status}`);
-        }
-
-        const data = await response.json();
-        if (data.documents && data.documents.length > 0) {
-          const formattedBooks = [];
-          const titlesSet = new Set();
-
-          data.documents.forEach((doc) => {
-            if (!titlesSet.has(doc.title)) {
-              titlesSet.add(doc.title);
-              formattedBooks.push({
-                title: doc.title,
-                imageUrl: doc.thumbnail,
-              });
-            }
-          });
-          setBooksData(formattedBooks);
-          setCurrentIndex(0); // 새로 검색하면 첫 번째 책부터 보여줘
-        } else {
-          console.log("검색 결과가 없습니다.");
-          setBooksData([]);
-        }
-      } catch (error) {
-        console.error("책 정보 API 호출 중 오류 발생:", error);
-        setBooksData([]);
-      }
+    const loadBooks = async () => {
+      const data = await fetchBooks(category);
+      setBooksData(data);
+      setCurrentIndex(0);
     };
-    fetchBooks();
-  }, [apiKey, query]);
+    loadBooks(); // 함수 호출
+  }, [category]);
+
+  // useEffect(() => {
+  //   const fetchBooks = async () => {
+  //     try {
+  //       const res = await axios.get(`/api/books?category=${category}`);
+  //       if (Array.isArray(res.data)) {
+  //         setBooksData(res.data); // 배열 그대로 저장
+  //         setCurrentIndex(0);
+  //       } else {
+  //         console.error("도서 데이터가 배열이 아님:", res.data);
+  //         setBooksData([]);
+  //       }
+  //     } catch (error) {
+  //       console.error("도서 목록 가져오기 실패:", error);
+  //       setBooksData([]);
+  //     }
+  //   };
+  //   fetchBooks();
+  // }, [category]);
 
   const handlePrev = () => {
-    setCurrentIndex((prevIndex) => Math.max(prevIndex - 6, 0)); // 0 아래로는 못 내려가게
+    setCurrentIndex((prev) => Math.max(prev - 6, 0));
   };
 
-  const handleNext = () => {
-    setCurrentIndex((prevIndex) =>
-      Math.min(prevIndex + 6, booksData.length - 6)
-    ); // 마지막 이상은 못 가게
-  };
+  const handleNext = () =>
+    setCurrentIndex((prev) => Math.min(prev + 6, booksData.length - 6));
 
   return (
     <div className="bookCarouselWrapper">
@@ -72,9 +60,9 @@ function BookCarousel({ query, title }) {
           {booksData
             .slice(currentIndex, currentIndex + 6)
             .map((book, index) => (
-              <div key={index} className="bookCoverWrapper">
+              <div key={book.id || index} className="bookCoverWrapper">
                 <img
-                  src={book.imageUrl}
+                  src={book.thumbnail || defaultThumbnail}
                   alt={book.title}
                   className="bookCoverImage"
                 />
@@ -83,9 +71,9 @@ function BookCarousel({ query, title }) {
             ))}
         </div>
         <button
-          className="nextIcon"
           onClick={handleNext}
-          disabled={currentIndex === booksData.length - 6}
+          disabled={currentIndex >= booksData.length - 6}
+          className="nextIcon"
         >
           <img src={iconRight} />
         </button>
