+++ frontend/src/pages/Home.jsx
@@ -28,14 +28,19 @@ function Home() {
       />
       <MainGrid />
       {/* BookCarousel 컴포넌트 사용 */}
-      <BookCarousel title="개발자가 찾는 추천도서" query={"개발" + "엔드"} />
+      {/* <BookCarousel title="개발자가 찾는 추천도서" query={"개발" + "엔드"} /> */}
       {/* 키워드 조합시 + 사용 */}
-      <BookCarousel title="한강 작가 특별전!" query={"한강" + "작가"} />
+      {/* <BookCarousel title="한강 작가 특별전!" query={"한강" + "작가"} /> */}
       {/* 키워드 조합시 + 사용 */}
-      <BookCarousel title="가장 많이찾는 경제도서" query={"세계" + "경제"} />
+      {/* <BookCarousel title="가장 많이찾는 경제도서" query={"세계" + "경제"} /> */}
       {/* 키워드 조합시 + 사용 */}
-      <BookCarousel title="봄을 맞이하며, 시집 추천해요" query={"시"} />
+      {/* <BookCarousel title="봄을 맞이하며, 시집 추천해요" query={"시"} /> */}
       {/* 키워드 조합시 + 사용  */}
+      {/* json-server의 books 엔드포인트에서 category 기준으로 데이터 조회 */}
+      <BookCarousel title="봄을 맞이하며, 시집 추천해요" category="시" />
+      <BookCarousel title="한강 작가 특별전" category="한강작가" />
+      <BookCarousel title="개발자가 찾는 추천도서" category="개발" />
+      <BookCarousel title="가장 많이 찾는 경제도서" category="세계경제" />
     </div>
   );
 }
